+++ client/src/router/index.js
@@ -1,5 +1,5 @@
 import {createRouter, createWebHistory } from 'vue-router';
-import OrderPage from "../pages/OrderPage.vue";
+import OrderPage from "../pages/client/ClientOrderPage.vue";
 import { useUserStore } from '@/stores/userStore';
 
 const requireEmployee = (to, from, next) => {
@@ -63,12 +63,12 @@ const routes = [
         component: () => import('@/pages/ProfileDataPage.vue'),
         meta: { requiresAuth: true }
       },
-        {
-            path: "/orders/:id",
-            name: "order",
-            component: OrderPage
-        }
-
+      {
+        path: "/orders/:id",
+        name: "order",
+        component: () => import('@/pages/OrderPage.vue'),
+        meta: { requiresAuth: true }
+      },
       {
         path: 'admin/users',
         name: 'AdminUsers',
@@ -85,14 +85,6 @@ const routes = [
     beforeEnter: requireEmployee,
     meta: { requiresAuth: true }
   },
-  
-  {
-    path: '/order/:id',  
-    name: 'OrderDetail',
-    component: () => import('@/pages/employee/OrderDetail.vue'),
-    beforeEnter: requireEmployee,  
-    meta: { requiresAuth: true }
-  }
 ];
 
 // Создание экземпляра маршрутизатора
