+++ server/src/controllers/orderController.js
@@ -16,10 +16,6 @@ export const createOrder = async (req, res) => {
 
 export const getOrders = async (req, res) => {
     try {
-        if (req.user.role !== 'admin' && req.user.role !== 'employee') {
-            return res.status(403).json({ message: 'Доступ запрещен' });
-        }
-
         const orders = await Order.find(req.query);
         res.json(orders);
     } catch (error) {
